backend/matching-engine/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-storage
# spacy
# sentence-transformers
# torch
pandas
db-dtypes
pytest
# transformers
redis[hiredis]
numpy
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace

